+++ ratelimit/keeper/genesis.go
@@ -1,4 +1,4 @@
-package ratelimit
+package keeper
 
 import (
 	"strconv"
@@ -6,13 +6,12 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/Stride-Labs/ibc-rate-limiting/ratelimit/keeper"
 	"github.com/Stride-Labs/ibc-rate-limiting/ratelimit/types"
 )
 
 // InitGenesis initializes the capability module's state from a provided genesis
 // state.
-func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
+func (k Keeper) InitGenesis(ctx sdk.Context, genState types.GenesisState) {
 	k.SetParams(ctx, genState.Params)
 	for _, rateLimit := range genState.RateLimits {
 		k.SetRateLimit(ctx, rateLimit)
@@ -38,7 +37,7 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 }
 
 // ExportGenesis returns the capability module's exported genesis.
-func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
+func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 	genesis := types.DefaultGenesis()
 
 	genesis.Params = k.GetParams(ctx)
